Added 2 API smoke tests
Created smoke tests for ComputeEngineAPI and UserNetworkAPI. I'm still working on DataStorageAPI because it seems to be a bit more complicated. I am likely missing a few things from these tests, so I will add them if I am. I also am currently not able to run the tests because of some issue with my file paths that I have to figure out.