Public Method Parameter Validation (#39)
* UserNetwork Parameter Validation

Added parameter validation to the UserNetworkAPI implementation. This checks the input and output locations. Additionally, the UserNetworkAPI Test had to be altered to not use the null locations. Before the change, it was failing due to the UserNetworkAPIImpl changes.

* ComputeEngine Parameter Validation

Added parameter validation for ComputeEngineAPI implementation. Most of the parameters were already being validated. Validation that the input is a positive integer was added.

* DataStorage Parameter Validation

Added parameter validation for DataStorageAPI implementation.

* Updated ComputeEngineAPIImpl

Added comments clearly describing the validation.

* Updated UserNetworkAPIImpl

Added clearer comments.

* Added Back Checkpoint 4 Feedback

The changes disappeared after merging.