Created CacheBenchmark
Created a simple time logging benchmark that will measure the time it takes to compute 10,000 integers with 100 repeated values before and after implementing a cache. This is done simply by clearing the cache before the first 10,000, measuring, then running another 10,000 computations and measuring again. The actual CachedComputeEngineAPIImpl still needs to be created, but this test should work as a benchmark to see if this is a bottleneck or not.